1.修改下容易产生问题的字段 "SensorUUID",这里容易在生成setter和getter的时候出现问题，呵呵呵 2.mybatis配置文件中添加几组alias

+++ src/main/java/com/ca/bms/dao/UserDao.java
@@ -29,7 +29,7 @@ public interface UserDao {
 	 * @param userid
 	 * @return
 	*/
-	UserEntity getUserById(int userid);
+	UserEntity getUserById(Long userid);
 	/**
 	 * 通过User查找User
 	 * @param user
